[heap] Simplify IncrementalMarking::StepSizeToMakeProgress.
This removes the ad-hoc ramp up period for the step size. The heuristic
was introduced before concurrent marking and is no longer useful.
The heuristic makes incremental marking progress very slow for short
running programs that do not allocate in the old generation.

Replace it with a simple cap on the step size in bytes.

Bug: v8:7253
Change-Id: If50fbd49e38135d006427f607c9a6d212c24fba7
Reviewed-on: https://chromium-review.googlesource.com/c/1288815
Reviewed-by: Hannes Payer <hpayer@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Ulan Degenbaev <ulan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#56807}