[wasm] Minor cleanup ExecuteCompilation for Turbofan
Two zones in ExecuteCompilation have the same scope, so we can combine
them.

In addition I moved measurement code at the end of the function so that
we don't have to store the result of the compilation in a helper.

R=clemensh@chromium.org

Bug: v8:8238
Cq-Include-Trybots: luci.chromium.try:linux_chromium_rel_ng
Change-Id: I3b4790fa62adbe361d41be188c7f0510b12d957c
Reviewed-on: https://chromium-review.googlesource.com/c/1291072
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Hammacher <clemensh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#56842}